Added progress spinner and button disable features
Now when the submit button is clicked on a form, a spinner with the same color as the button's background appears to show the user that the form is submitting. Also the submit button is disabled to prevent additional submissions.

The spinner scales based on the button font size. Since that is how the button resizes, the spinner maintains it's size ratio with the submit button.